Update `docker compose up` strategy.
Create a new `web.Dockerfile`. This time were sure to add our app code in both our app and web containers. I'm still not sure the best way to include repeated add steps like this.

This commit also makes sure that all the added files are owned by the correct user and group. We then are sure to set the php-fpm worker and web worker to use this same user.

Finally, we get the vips thumbnailer working.

At this phase Omeka S is up and seemingly functioning as normal.