Add sideload volume and ingest `./src/sideload`
I've added the `ingest` service to `compose.yaml`.

This can act as a one time file ingester to populate a volume with src data.

Use the `ingest.Dockerfile` to add the files to the mounted directory. This then populates the volume.

Afterwards, the container will exit and it and the image are safe to delete.

When you don't need to ingest data or files, you can comment out the `ingest` service entirely.